Extract multiquery (#94)
As there remain several important questions to be answered about some MultiQuery functionality, this PR will move these methods into a separate module to prevent too much of the code getting entangled again until we have something that will be maintained long term.

```python
client.read_multi(...)
```

becomes

```python
from bfabric.experimental.multi_query import MultiQuery
MultiQuery(client).read_multi(...)
```
