+++ bfabric/bfabric.py
@@ -18,7 +18,6 @@ import importlib.metadata
 import logging
 import os
 from contextlib import contextmanager
-from copy import deepcopy
 from datetime import datetime
 from enum import Enum
 from pprint import pprint
@@ -32,7 +31,7 @@ from bfabric.cli_formatting import HostnameHighlighter, DEFAULT_THEME
 from bfabric.engine.engine_suds import EngineSUDS
 from bfabric.engine.engine_zeep import EngineZeep
 from bfabric.results.result_container import ResultContainer
-from bfabric.utils.paginator import compute_requested_pages, BFABRIC_QUERY_LIMIT, page_iter
+from bfabric.utils.paginator import compute_requested_pages, BFABRIC_QUERY_LIMIT
 
 
 class BfabricAPIEngineType(Enum):
@@ -209,6 +208,23 @@ class Bfabric:
             results.assert_success()
         return results
 
+    def exists(
+        self, endpoint: str, key: str, value: int | str, query: dict[str, Any] | None = None, check: bool = True
+    ) -> bool:
+        """Returns whether an object with the specified key-value pair exists in the specified endpoint.
+        Further conditions can be specified in the query.
+        :param endpoint: the endpoint to check, e.g. "sample"
+        :param key: the key to check, e.g. "id"
+        :param value: the value to check, e.g. 123
+        :param query: additional query conditions (optional)
+        :param check: whether to raise an error if the response is not successful
+        """
+        query = query or {}
+        results = self.read(
+            endpoint=endpoint, obj={**query, key: value}, max_results=1, check=check, return_id_only=True
+        )
+        return len(results) > 0
+
     def upload_resource(
         self, resource_name: str, content: bytes, workunit_id: int, check: bool = True
     ) -> ResultContainer:
@@ -230,113 +246,6 @@ class Bfabric:
             },
             check=check,
         )
-
-    ############################
-    # Multi-query functionality
-    ############################
-
-    # TODO: Is this scope sufficient? Is there ever more than one multi-query parameter, and/or not at the root of dict?
-    def read_multi(
-        self,
-        endpoint: str,
-        obj: dict,
-        multi_query_key: str,
-        multi_query_vals: list,
-        readid: bool = False,
-        idonly: bool = False,
-    ) -> ResultContainer:
-        """
-        Makes a 1-parameter multi-query (there is 1 parameter that takes a list of values)
-        Since the API only allows BFABRIC_QUERY_LIMIT queries per page, split the list into chunks before querying
-        :param endpoint: endpoint
-        :param obj: query dictionary
-        :param multi_query_key:  key for which the multi-query is performed
-        :param multi_query_vals: list of values for which the multi-query is performed
-        :param readid: whether to use reading by ID. Currently only available for engine=SUDS
-            TODO: Test the extent to which this method works. Add safeguards
-        :param idonly: whether to return only the ids of the objects
-        :return: List of responses, packaged in the results container
-
-        NOTE: It is assumed that there is only 1 response for each value.
-        """
-
-        response_tot = ResultContainer([], total_pages_api=0)
-        obj_extended = deepcopy(obj)  # Make a copy of the query, not to make edits to the argument
-
-        # Iterate over request chunks that fit into a single API page
-        for page_vals in page_iter(multi_query_vals):
-            obj_extended[multi_query_key] = page_vals
-
-            # TODO: Test what happens if there are multiple responses to each of the individual queries.
-            #     * What would happen?
-            #     * What would happen if total number of responses would exceed 100 now?
-            #     * What would happen if we naively made a multi-query with more than 100 values? Would API paginate
-            #       automatically? If yes, perhaps we don't need this method at all?
-            # TODO: It is assumed that a user requesting multi_query always wants all of the pages. Can anybody think of
-            #   exceptions to this?
-            response_this = self.read(endpoint, obj_extended, max_results=None, return_id_only=idonly)
-            response_tot.extend(response_this)
-
-        return response_tot
-
-    # NOTE: Save-multi method is likely useless. When saving multiple objects, they all have different fields.
-    #    One option would be to provide a dataframe, but it might struggle with nested dicts
-    #    Likely best solution is to not provide this method, and let users run a for-loop themselves.
-    # def save_multi(self, endpoint: str, obj_lst: list, **kwargs) -> ResultContainer:
-    #     response_tot = ResultContainer([], self.result_type, total_pages_api = 0)
-    #
-    #     # Iterate over request chunks that fit into a single API page
-    #     for page_objs in page_iter(obj_lst):
-    #         response_page = self.save(endpoint, page_objs, **kwargs)
-    #         response_tot.extend(response_page)
-    #
-    #     return response_tot
-
-    def delete_multi(self, endpoint: str, id_list: list) -> ResultContainer:
-        response_tot = ResultContainer([], total_pages_api=0)
-
-        if len(id_list) == 0:
-            print("Warning, empty list provided for deletion, ignoring")
-            return response_tot
-
-        # Iterate over request chunks that fit into a single API page
-        for page_ids in page_iter(id_list):
-            response_page = self.delete(endpoint, page_ids)
-            response_tot.extend(response_page)
-
-        return response_tot
-
-    def exists(self, endpoint: str, key: str, value: list[int | str] | int | str) -> bool | list[bool]:
-        """
-        :param endpoint:  endpoint
-        :param key:       A key for the query (e.g. id or name)
-        :param value:     A value or a list of values
-        :return:          Return a single bool or a list of bools for each value
-            For each value, test if a key with that value is found in the API.
-        """
-        is_scalar = isinstance(value, (int, str))
-
-        # 1. Read data for this id
-        if is_scalar:
-            results = self.read(endpoint, {key: value})
-        elif isinstance(value, list):
-            results = self.read_multi(endpoint, {}, key, value)
-        else:
-            raise ValueError("Unexpected data type", type(value))
-
-        # 2. Extract all the ids for which there was a response
-        result_vals = []
-        for r in results.results:
-            if key in r:
-                result_vals += [r[key]]
-            elif "_" + key in r:  # TODO: Remove this if SUDS bug is ever resolved
-                result_vals += [r["_" + key]]
-
-        # 3. For each of the requested ids, return true if there was a response and false if there was not
-        if is_scalar:
-            return value in result_vals
-        else:
-            return [val in result_vals for val in value]
 
     def get_version_message(self) -> str:
         """Returns the version message as a string."""
